Fix animation to spec
