+++ composables/src/debug/java/com/google/android/horologist/composables/MarqueeTextPreview.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.composables
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(
+    widthDp = 100,
+    heightDp = 20,
+    backgroundColor = 0xFF000000,
+    showBackground = true
+)
+@Composable
+fun MarqueeTextPreview() {
+    MarqueeText(
+        text = "A very long text strings",
+        modifier = Modifier
+            .fillMaxWidth(),
+        textAlign = TextAlign.Center,
+    )
+}
